feat: implemented measurement by sensor's interval with the generation of random temperature and humidity (based on the current ones)

+++ src/main/java/fr/miage/acm/measurementservice/api/ApiSensor.java
@@ -16,8 +16,8 @@ public class ApiSensor {
     private DeviceState state;
     private float interval;
     private Field field;
-    private Integer currentTemperature;
-    private Integer currentHumidity;
+    private Float currentTemperature;
+    private Float currentHumidity;
 
     public ApiSensor(Sensor sensor) {
         this.id = sensor.getId();
